backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
alembic==1.14.1
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
anthropic==0.45.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
